Part 1 (SQL) of the Twitch Visualization Project
As the Codecademy README indicates, we were given a scrubbed dataset of stream and chat data from Twitch to explore. We practiced our SQL queries to answer a series of questions about this data.

The SQL database file is too large to push to GitHub. Please contact me if you would like to see a copy of the database file.