Part 1 of Twitch Visualization Project
Twitch provided a scrubbed dataset containing stream and chat data for Codecademy students to explore. Part 1 of this project allowed us to practice our SQL skills to access and query the datasets.

The SQL database file was too large to commit to GitHub. If you are interested in looking at the database file, please contact me.