Twitter Classification Project
A two-part project to practice our skills on datasets of Tweets using two supervised machine learning classifiers we learned recently learned about.

In the first part, we use the K-Nearest-Neighbors classifier (from the sklearn library) and in the second part, we use the Naive-Bayes classifier (also from the sklearn library).

The project was completed on Jupyter notebooks.